+++ CHANGELOG.md
@@ -1,4 +1,14 @@
-# HTMLFilter 1.0 - CHANGELOG
+# HTMLFilter 2.0 - CHANGELOG
+
+**version 2.0 (09/03/2019)**
+- Fix error message when clicked target is null.
+- Create Settings page.
+- Add new icons.
+- Add internationalization with i18n API.
+- Add French, Italian and Spannish translations.
+- Add a new option to enable/disable HTMLFilter in page context menu (no Firefox Android support).
+- Add Settings to browserAction context menu for Firefox 53+.
+- Code improvements.
 
 **version 1.0 (01/11/2018)**
 - First version :)